+++ src/Service/ImageUploadService.php
@@ -1,0 +1,39 @@
+<?php
+
+namespace App\Service;
+
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\String\Slugger\SluggerInterface;
+use Symfony\Component\HttpFoundation\File\Exception\FileException;
+
+class ImageUploadService extends AbstractController
+{
+    private $slugger;
+
+    public function __construct(SluggerInterface $slugger)
+    {
+        $this->slugger = $slugger;
+    }
+    
+    public function ImageUpload($form, $chambre)
+    {
+        $imageFile = $form->get('image')->getData();       
+        if ($imageFile) {
+            $originalFilename = pathinfo($imageFile->getClientOriginalName(), PATHINFO_FILENAME);
+            $safeFilename = $this->slugger->slug($originalFilename);
+
+            $newFilename = $safeFilename.'-'.uniqid().'.'.$imageFile->guessExtension();
+
+            try {
+                $imageFile->move(
+                    $this->getParameter('img_upload'),
+                    $newFilename
+                );
+            } catch (FileException $e) {
+
+            }
+            $chambre->setImage($newFilename);
+
+        }
+    }
+}
